requirements.txt
numpy==1.23.5
Pillow==9.3.0
streamlit==1.15.2
